Move EmptyIfBlockInspectionTests for QuickFixes into relevant files under RD.QuickFixes
